requirements.txt
certifi==2024.2.2
charset-normalizer==3.3.2
idna==3.7
joblib==1.4.2
numpy==1.26.4
pandas==2.2.2
pillow==10.3.0
psutil==5.9.8
pydub==0.25.1
python-dateutil==2.9.0.post0
pytz==2024.1
requests==2.32.2
scikit-learn==1.5.0
scipy==1.13.1
six==1.16.0
threadpoolctl==3.5.0
tzdata==2024.1
urllib3==2.2.1
